blend index.js with readme.md
