--- index.js
@@ -1,16 +1,0 @@
-const fromObs = observable => (start, sink) => {
-  if (start !== 0) return;
-  let dispose;
-  sink(0, t => {
-    if (t === 2 && dispose) {
-      dispose();
-    }
-  });
-  dispose = observable.subscribe({
-    next: x => sink(1, x),
-    error: e => sink(2, e),
-    complete: () => sink(2)
-  });
-};
-
-module.exports = fromObs;
